requirements.txt
aiohttp
aiohttp_cors
aiopg
asyncio
black
gunicorn
isort
psycopg2-binary
pyyaml
yoyo-migrations